requirements.txt
flask==3.0.3
python-dotenv==1.0.1
openai==1.37.0
faiss-cpu==1.8.0
numpy==1.26.4
